requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

kafka-python==2.0.2

psycopg2-binary==2.9.9

redis==5.0.1